Move aggregates in sql to own file and apply formatting
